fix infinite render
